Move useStep hook
